+++ src/utils/services.py
@@ -1,14 +1,23 @@
 from __future__ import annotations
 
 import os
+from dataclasses import dataclass
 from typing import List
+from typing import Optional
 
+from configs.service_config import ServiceConfig
 from exceptions import ConfigNotFoundError
 from exceptions import ConfigParseError
 from exceptions import ConfigValidationError
 from exceptions import ServiceNotFoundError
 from utils.devenv import get_coderoot
-from utils.types import Service
+
+
+@dataclass
+class Service:
+    name: str
+    repo_path: str
+    config: ServiceConfig
 
 
 def get_local_services(coderoot: str) -> List[Service]:
@@ -27,14 +36,25 @@ def get_local_services(coderoot: str) -> List[Service]:
             Service(
                 name=service_name,
                 repo_path=repo_path,
-                service_config=service_config,
+                config=service_config,
             )
         )
     return services
 
 
-def find_matching_service(service_name: str) -> Service:
+def find_matching_service(service_name: Optional[str]) -> Service:
     """Find a service with the given name."""
+    if service_name is None:
+        from configs.service_config import load_service_config_from_file
+
+        repo_path = os.getcwd()
+        service_config = load_service_config_from_file(repo_path)
+
+        return Service(
+            name=service_config.service_name,
+            repo_path=repo_path,
+            config=service_config,
+        )
     coderoot = get_coderoot()
     services = get_local_services(coderoot)
     for service in services:
